requirements.txt
requests
PyYAML
feedparser
python-dotenv